Solved 07b

+++ src/intcode_computer.rs
@@ -4,12 +4,12 @@ use std::process::exit;
 
 pub struct IntcodeComputer {
     instructions: Map<usize, Instruction>,
-    memory: Vec<i32>,
+    memory: Vec<i64>,
     state: State,
 }
 
 impl IntcodeComputer {
-    pub fn new(program: &[i32], input: Receiver<i32>, output: Sender<i32>, debug: bool) -> IntcodeComputer {
+    pub fn new(program: &[i64], input: Receiver<i64>, output: Sender<i64>, debug: bool) -> IntcodeComputer {
         let mut memory = vec![0; program.len()];
         memory.copy_from_slice(&program[..]);
 
@@ -27,7 +27,7 @@ impl IntcodeComputer {
         }
     }
     pub fn step(&mut self) {
-        let mem: &mut [i32] = self.memory.as_mut();
+        let mem: &mut [i64] = self.memory.as_mut();
         self.state.param_modes = decode_params(mem[self.state.ip]);
         if let Some(ins) = self.instructions.get(decode_instruction(mem[self.state.ip])) {
             if self.state.debug {
@@ -47,7 +47,7 @@ impl IntcodeComputer {
     }
 }
 
-fn get<'a>(state: &State, mem: &'a mut [i32], pos: usize) -> &'a mut i32 {
+fn get<'a>(state: &State, mem: &'a mut [i64], pos: usize) -> &'a mut i64 {
     if state.param_modes[pos - 1] {
         mem.get_mut(state.ip + pos).unwrap()
     } else {
@@ -55,13 +55,13 @@ fn get<'a>(state: &State, mem: &'a mut [i32], pos: usize) -> &'a mut i32 {
     }
 }
 
-fn decode_params(value: i32) -> [bool; 3] {
+fn decode_params(value: i64) -> [bool; 3] {
     let string = value.to_string();
     let mut chars = string.chars();
     [chars.nth_back(2).unwrap_or('0') == '1', chars.next_back().unwrap_or('0') == '1', chars.next_back().unwrap_or('0') == '1']
 }
 
-fn decode_instruction(value: i32) -> usize {
+fn decode_instruction(value: i64) -> usize {
     (value % 100) as usize
 }
 
@@ -69,15 +69,15 @@ struct State {
     param_modes: [bool; 3],
     ip: usize,
     halt: bool,
-    input: Receiver<i32>,
-    output: Sender<i32>,
+    input: Receiver<i64>,
+    output: Sender<i64>,
     debug: bool,
 }
 
 struct Instruction {
     name: &'static str,
     ip_change: usize,
-    exec: fn(state: &mut State, mem: &mut [i32]),
+    exec: fn(state: &mut State, mem: &mut [i64]),
 }
 
 fn fill_instructions() -> Map<usize, Instruction> {
@@ -85,42 +85,42 @@ fn fill_instructions() -> Map<usize, Instruction> {
     map.insert(99, Instruction {
         name: "HALT",
         ip_change: 1,
-        exec: |state: &mut State, _: &mut [i32]| {
+        exec: |state: &mut State, _: &mut [i64]| {
             state.halt = true;
         },
     });
     map.insert(1, Instruction {
         name: "ADD",
         ip_change: 4,
-        exec: |state: &mut State, mem: &mut [i32]| {
+        exec: |state: &mut State, mem: &mut [i64]| {
             *get(&state, mem, 3) = *get(&state, mem, 1) + *get(&state, mem, 2);
         },
     });
     map.insert(2, Instruction {
         name: "MUL",
         ip_change: 4,
-        exec: |state: &mut State, mem: &mut [i32]| {
+        exec: |state: &mut State, mem: &mut [i64]| {
             *get(&state, mem, 3) = *get(&state, mem, 1) * *get(&state, mem, 2);
         },
     });
     map.insert(3, Instruction {
         name: "IN",
         ip_change: 2,
-        exec: |state: &mut State, mem: &mut [i32]| {
+        exec: |state: &mut State, mem: &mut [i64]| {
             *get(&state, mem, 1) = state.input.recv().unwrap();
         },
     });
     map.insert(4, Instruction {
         name: "OUT",
         ip_change: 2,
-        exec: |state: &mut State, mem: &mut [i32]| {
+        exec: |state: &mut State, mem: &mut [i64]| {
             let _ = state.output.send(*get(&state, mem, 1));
         },
     });
     map.insert(5, Instruction {
         name: "JNZ",
         ip_change: 3,
-        exec: |state: &mut State, mem: &mut [i32]| {
+        exec: |state: &mut State, mem: &mut [i64]| {
             if *get(&state, mem, 1) != 0 {
                 state.ip = *get(&state, mem, 2) as usize - 3;
             }
@@ -129,7 +129,7 @@ fn fill_instructions() -> Map<usize, Instruction> {
     map.insert(6, Instruction {
         name: "JZ",
         ip_change: 3,
-        exec: |state: &mut State, mem: &mut [i32]| {
+        exec: |state: &mut State, mem: &mut [i64]| {
             if *get(&state, mem, 1) == 0 {
                 state.ip = *get(&state, mem, 2) as usize - 3;
             }
@@ -138,7 +138,7 @@ fn fill_instructions() -> Map<usize, Instruction> {
     map.insert(7, Instruction {
         name: "LESS",
         ip_change: 4,
-        exec: |state: &mut State, mem: &mut [i32]| {
+        exec: |state: &mut State, mem: &mut [i64]| {
             *get(&state, mem, 3) = if *get(&state, mem, 1) < *get(&state, mem, 2) {
                 1
             } else {
@@ -149,7 +149,7 @@ fn fill_instructions() -> Map<usize, Instruction> {
     map.insert(8, Instruction {
         name: "EQUALS",
         ip_change: 4,
-        exec: |state: &mut State, mem: &mut [i32]| {
+        exec: |state: &mut State, mem: &mut [i64]| {
             *get(&state, mem, 3) = if *get(&state, mem, 1) == *get(&state, mem, 2) {
                 1
             } else {
